requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
pycryptodome

